[PATCH] Make arch-independent syscalls return long
From: "Randy.Dunlap" <randy.dunlap@verizon.net>

Fix up various syscalls to return longs, as x86_64 and ia64 (at least)
require.

+++ include/linux/mm.h
@@ -421,7 +421,7 @@ extern int install_page(struct mm_struct *mm, struct vm_area_struct *vma, unsign
 extern int handle_mm_fault(struct mm_struct *mm,struct vm_area_struct *vma, unsigned long address, int write_access);
 extern int make_pages_present(unsigned long addr, unsigned long end);
 extern int access_process_vm(struct task_struct *tsk, unsigned long addr, void *buf, int len, int write);
-extern int sys_remap_file_pages(unsigned long start, unsigned long size, unsigned long prot, unsigned long pgoff, unsigned long nonblock);
+extern long sys_remap_file_pages(unsigned long start, unsigned long size, unsigned long prot, unsigned long pgoff, unsigned long nonblock);
 
 
 int get_user_pages(struct task_struct *tsk, struct mm_struct *mm, unsigned long start,
